schwarz-precond: remove IdentityOperator from public API
It existed only as test infrastructure for the now-deleted CG dedup
path. Now defined as a private struct inside lsmr/tests.rs (its sole
legitimate use), with the three trivial integration tests deleted.
